requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
opencv-python>=4.7.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0

# Optional: For faster training
# accelerate>=0.20.0
# bitsandbytes>=0.39.0
